requirements.txt
pyyaml
pydantic
paho-mqtt
umodbus